requirements.txt
asgiref==3.5.2
beautifulsoup4==4.11.1
certifi==2022.6.15
charset-normalizer==2.0.12
Django==4.0.5
django-cleanup==6.0.0
django-cors-headers==3.13.0
django-rest-framework==0.1.0
django-split-settings==1.1.0
djangorestframework==3.13.1
ffmpeg-python==0.2.0
future==0.18.2
idna==3.3
Pillow==9.2.0
pydub==0.25.1
pytz==2022.1
requests==2.28.0
six==1.16.0
soupsieve==2.3.2.post1
SpeechRecognition==3.8.1
sqlparse==0.4.2
urllib3==1.26.9
